+++ tests/WebHemiTest/TestService/TestFalseValidator.php
@@ -11,23 +11,33 @@
  */
 namespace WebHemiTest\TestService;
 
-use WebHemi\Validator\ServiceInterface as ValidatorInterface;
+use WebHemi\Validator\ValidatorInterface;
 
 /**
- * Class TestTrueValidator.
+ * Class TestFalseValidator.
  */
 class TestFalseValidator implements ValidatorInterface
 {
     /**
      * Validates data.
      *
-     * @param mixed $data
+     * @param array $values
      * @return boolean
      */
-    public function validate($data) : bool
+    public function validate(array $values) : bool
     {
-        unset($data);
+        unset($values);
         return false;
+    }
+
+    /**
+     * Retrieve valid data.
+     *
+     * @return array
+     */
+    public function getValidData() : array
+    {
+        return [];
     }
 
     /**